fastapi-strawberry-example/requirements.txt
uvicorn[standard]==0.17.6
fastapi[all]==0.78.0
strawberry-graphql[fastapi]==0.122.1
websockets==10.3